pyproject.toml
[project]
name = "gavin-the-fish"
version = "0.1.0"
description = "API to expose Raycast scripts as webhook endpoints"
authors = [
    { name = "Caleb", email = "caleb@example.com" }
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=0.19.0",
    "playwright>=1.40.0",
    "httpx>=0.24.0",
    "rich>=13.7.0",
    "rumps>=0.4.0",
    "elevenlabs>=0.2.26",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gavin_the_fish"]
